[REF] web,*: convert hotkey service to plugin
*: account,analytic,barcodes,html_builder,html_editor,mail,product,stock,survey,
website,website_blog

In this commit, we rewrite the hotkey service as a plugin.
For legacy purposes, we keep the hotkey service (as a service).
