+++ addons/account/static/src/components/product_catalog/search/search_panel/search_panel.js
@@ -1,6 +1,6 @@
 import { onWillStart, proxy, signal } from "@odoo/owl";
 import { formatCurrency } from "@web/core/currency";
-import { getActiveHotkey } from "@web/core/hotkeys/hotkey_service";
+import { getActiveHotkey } from "@web/core/hotkeys/hotkey_utils";
 import { useBus, useService } from "@web/core/utils/hooks";
 import { useNestedSortable } from "@web/core/utils/nested_sortable";
 import { useSubEnv } from "@web/owl2/utils";
