+++ addons/account/static/src/components/product_catalog/search/section_row/section_row.js
@@ -1,7 +1,7 @@
 import { Component, onMounted, onPatched, signal, t, useProps } from "@odoo/owl";
 import { Dropdown } from "@web/core/dropdown/dropdown";
 import { DropdownItem } from "@web/core/dropdown/dropdown_item";
-import { getActiveHotkey } from "@web/core/hotkeys/hotkey_service";
+import { getActiveHotkey } from "@web/core/hotkeys/hotkey_utils";
 
 export class SectionRow extends Component {
     static template = "account.SectionRow";
