+++ addons/html_builder/static/src/utils/keyboard_navigation.js
@@ -1,6 +1,6 @@
 import { onMounted, onPatched } from "@odoo/owl";
 import { browser } from "@web/core/browser/browser";
-import { getActiveHotkey } from "@web/core/hotkeys/hotkey_service";
+import { getActiveHotkey } from "@web/core/hotkeys/hotkey_utils";
 import { localization } from "@web/core/l10n/localization";
 
 /**
