+++ addons/web/static/src/core/autocomplete/autocomplete.js
@@ -1,7 +1,7 @@
 import { useAutofocus, useService } from "@web/core/utils/hooks";
 import { isScrollableY, scrollTo } from "@web/core/utils/scrolling";
 import { useDebounced } from "@web/core/utils/timing";
-import { getActiveHotkey } from "@web/core/hotkeys/hotkey_service";
+import { getActiveHotkey } from "@web/core/hotkeys/hotkey_utils";
 import { usePosition } from "@web/core/position/position_hook";
 import { Component, onWillUpdateProps, proxy, signal, t, useListener, useProps } from "@odoo/owl";
 import { mergeClasses } from "@web/core/utils/classname";
