+++ addons/web/static/src/views/fields/input_field_hook.js
@@ -1,5 +1,5 @@
 import { onMounted, onPatched, untrack, useListener, useProps } from "@odoo/owl";
-import { getActiveHotkey } from "@web/core/hotkeys/hotkey_service";
+import { getActiveHotkey } from "@web/core/hotkeys/hotkey_utils";
 import { useBus } from "@web/core/utils/hooks";
 import { onWillRender } from "@web/owl2/utils";
