+++ addons/website_blog/static/src/interactions/website_blog.js
@@ -4,7 +4,7 @@ import { registry } from "@web/core/registry";
 import { browser } from "@web/core/browser/browser";
 import { _t } from "@web/core/l10n/translation";
 import { verifyHttpsUrl } from "@website/utils/misc";
-import { getActiveHotkey } from "@web/core/hotkeys/hotkey_service";
+import { getActiveHotkey } from "@web/core/hotkeys/hotkey_utils";
 import { BlogNavSheet } from "./components/blog_nav_sheet";
 
 export class WebsiteBlog extends Interaction {
